+++ src/main/java/com/example/getmesocialservicedemo/resource/PhotoResource.java
@@ -1,11 +1,15 @@
 package com.example.getmesocialservicedemo.resource;
 
+import com.example.getmesocialservicedemo.model.FirebaseUser;
 import com.example.getmesocialservicedemo.model.Photo;
+import com.example.getmesocialservicedemo.service.FirebaseService;
 import com.example.getmesocialservicedemo.service.PhotoService;
+import com.google.firebase.auth.FirebaseAuthException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.io.IOException;
 import java.util.List;
 
 @RestController
@@ -14,6 +18,9 @@ public class PhotoResource {
 
     @Autowired
     private PhotoService photoService;
+
+    @Autowired
+    private FirebaseService firebaseService;
 
     @GetMapping
     public List<Photo> getAllPhotos(){
@@ -26,17 +33,28 @@ public class PhotoResource {
     }
 
     @PutMapping
-    public Photo updatePhoto(@RequestBody Photo photo){
-        return photoService.updatePhoto(photo);
+    public Photo updatePhoto(@RequestBody Photo photo, @RequestHeader(name = "idToken") String idToken) throws IOException, FirebaseAuthException {
+        FirebaseUser firebaseUser = firebaseService.authenticate(idToken);
+        if (firebaseUser!=null) {
+            return photoService.updatePhoto(photo);
+        }
+        return null;
     }
 
     @PostMapping
-    public Photo savePhoto(@Valid Photo photo){
-        return photoService.savePhoto(photo);
+    public Photo savePhoto(@Valid Photo photo, @RequestHeader(name = "idToken") String idToken) throws IOException, FirebaseAuthException{
+        FirebaseUser firebaseUser = firebaseService.authenticate(idToken);
+        if (firebaseUser!=null) {
+            return photoService.savePhoto(photo);
+        }
+        return null;
     }
 
     @DeleteMapping
-    public void deletePhoto(@RequestParam(name = "photoId") String photoId){
-        photoService.deletePhoto(photoId);
+    public void deletePhoto(@RequestParam(name = "photoId") String photoId, @RequestHeader(name = "idToken") String idToken) throws IOException, FirebaseAuthException{
+        FirebaseUser firebaseUser = firebaseService.authenticate(idToken);
+        if (firebaseUser!=null) {
+            photoService.deletePhoto(photoId);
+        }
     }
 }
